fixing the oauth issues caused by a bad redirect from github
