+++ package.json
@@ -32,6 +32,7 @@
   "dependencies": {
     "body-parser": "^1.13.1",
     "connect-mongo": "^0.8.1",
+    "cookie-parser": "^1.3.5",
     "cookie-session": "^1.1.0",
     "express": "^4.13.0",
     "express-session": "^1.11.3",
